requirements.txt
scrapy
libsql-experimental
python-dotenv